Move IngestJobValidatedEvent to core module
