+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/core/job/status/InMemoryIngestJobStatusStore.java
@@ -23,6 +23,7 @@ import sleeper.core.tracker.ingest.job.update.IngestJobAddedFilesEvent;
 import sleeper.core.tracker.ingest.job.update.IngestJobFailedEvent;
 import sleeper.core.tracker.ingest.job.update.IngestJobFinishedEvent;
 import sleeper.core.tracker.ingest.job.update.IngestJobStartedEvent;
+import sleeper.core.tracker.ingest.job.update.IngestJobValidatedEvent;
 
 import java.time.Instant;
 import java.util.ArrayList;
